changed login to 1 component only!!!!
