+++ african-marketplace/src/page/about.js
@@ -4,6 +4,9 @@ export default function AboutPage(props) {
   return (
     <div>
       <h2>About Page</h2>
+      <p>stuffffffffffflflflflflflfllflflflf
+        fkfkfkfkfkfkkfkfkfkfjhskksfjhsfhfksjf
+      </p>
     </div>
   );
 }
