+++ african-marketplace/src/page/home.js
@@ -1,5 +1,5 @@
 import React from "react"
-import FormLogin from "../component/FormLogin"
+import Login from './login'
 import ButtonStyle from "../css/Button"
 import Button from "../css/Button"
 import "../css/Home.css"
